Refactor into separate classes
We now have an

-Event handling class
-State machine class
-User input handler class

Additionally I have simplified things a bit by making everything
except debugee standard output handling execute on the main thread.
Since processing debugger commands (i.e. "backtrace" or whatever) is
synchronous with the target being stopped, we can wait until the
target is suspended before asking for user input, and if the target is
not suspended, we don't need another thread running to handle user
input.  This may change but the simplification is a better way to
proceed at the beginning until we know we need more complexity.