Move classindex to interface and replace with Hashmap
Also we now hash every substring of length >=3 of every class
name and reutrn that list on a hit, instead of using a Treemap and
navigating from the first hit through it's linked list of values to
find all matches. Index size is about 2.7MB in memory, which is only
slightly larger than the Treemap size (2.5MB) on a test case using Fibonacci.